rename index_feature to index_features (#295)
