+++ src/index_feature_list.h
@@ -18,7 +18,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "vsag/index_feature.h"
+#include "vsag/index_features.h"
 
 namespace vsag {
 class IndexFeatureList {
